Move handleRules into Rules module
